Figure out GC in Blocks
The current approach does not work when a block that generated block
gets out of scope. The solution is to store references to variables
introduced to the current scope (arguments and new allocations) on the
returned block. Since the caller will keep the returned block alive,
the caller will also keep the objects in the blocks' scope
alive. Blocks must also keep reference to the current block they are
housed in, so that if the caller loses reference to the generating
block, the generated blocked can still make use of all variables that
were available to the parents' scope.

NB: This is yet to be implemented in the generator.